requirements.txt
# torch
# torchvision
numpy
PyYAML
tqdm
test_tube
matplotlib
scipy
scikit-learn
scikit-image